refactor: split GridItem tests into grid-item.test.tsx (T008)
